Refactor linear solver files for future extension and consistent with other linalg files
